pyproject.toml
[project]
name = "flight-booking-agent"
version = "0.1.0"
description = "AI agent that books flights using LangGraph and Groq"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.6.6",
    "mcp>=1.0.0",
    "pybreaker>=1.2.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.30.0",
]
